Copy over dump of sourceforge bzip
I'm not currently super familiar with how transitions from Mercurial to
Git work, but it looks like a bunch of files were missing?

Repeated attempts to compile this library from the git repository failed
because of missing files and configs. Copying the contents of bzip file
for version 0.10 seemed to resolve these issues for me.